Make command processing concurrent
pigeon does not lose any user sent commands anymore.
this is a huge gain over its previouse behaviour.
