+++ pigeon.rb
@@ -17,3 +17,4 @@ require 'server'
 require 'crawler'
 require 'tasks'
 require 'command_task'
+require 'command_processor'
