+++ src/tasks.rb
@@ -1,24 +1,37 @@
+
 class Tasks
 
   def initialize
     @tasks = []
+    @mutex = Mutex.new
+    @resource = ConditionVariable.new
+    @has_data = false
   end
 
   def append(task)
-    @tasks << task
-  end
-
-  def count
-    @tasks.count
+    @mutex.synchronize do
+      @tasks << task
+      @has_data = true
+      @resource.signal
+    end
   end
 
   # returns oldest task stored on the task list.
   def deque
-    @tasks.shift unless empty?
+    @mutex.synchronize do
+      while @has_data == false
+        @resource.wait(@mutex)
+      end
+      @has_data = false if @tasks.count == 1
+      @resource.signal
+      @tasks.shift
+    end
   end
 
   def empty?
-    @tasks.empty?
+    @mutex.synchronize do
+      @tasks.empty?
+    end
   end
 
 end
